+++ src/main/java/com/blecua84/moneytransfers/services/models/Transfer.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -27,9 +28,9 @@ public class Transfer {
     private Account to;
 
     @Column(name = "amount")
-    private float amount;
+    private BigDecimal amount;
 
-    public Transfer(Account from, Account to, float amount) {
+    public Transfer(Account from, Account to, BigDecimal amount) {
         this.from = from;
         this.to = to;
         this.amount = amount;
